Implement IFluentControlStyle interface for custom CSS styling in Fluent components
