+++ src/Core/Components/TextInput/FluentTextInput.razor.cs
@@ -13,14 +13,8 @@ namespace Microsoft.FluentUI.AspNetCore.Components;
 /// <summary>
 /// A text input component that allows users to enter and edit a single line of text.
 /// </summary>
-public partial class FluentTextInput : FluentInputImmediateBase<string?>, IFluentComponentElementBase, ITooltipComponent, IFluentComponentChangeAfterKeyPress
+public partial class FluentTextInput : FluentInputImmediateBase<string?>, IFluentComponentElementBase, ITooltipComponent, IFluentComponentChangeAfterKeyPress, IFluentControlStyle
 {
-    /// <summary>
-    /// Tracks the last <see cref="ControlStyle"/> value applied to the internal input element, to avoid
-    /// unnecessary JS interop calls on every render.
-    /// </summary>
-    private string? _appliedControlStyle;
-
     /// <summary>
     /// Initializes a new instance of the <see cref="FluentTextInput"/> class.
     /// </summary>
@@ -145,13 +139,7 @@ public partial class FluentTextInput : FluentInputImmediateBase<string?>, IFluen
     [Parameter]
     public string? Width { get; set; }
 
-    /// <summary>
-    /// Gets or sets custom CSS applied to the internal <c>input</c> element (the shadow DOM "control" part).
-    /// Unlike <see cref="FluentComponentBase.Style"/> (applied to the host element), this is injected into the
-    /// component's shadow root, so it can also target pseudo-elements that <c>::part(control)</c> can't reach
-    /// from outside (e.g. <c>ControlStyle="::-ms-reveal { display: none; }"</c> to hide the Edge password reveal icon).
-    /// Plain declarations without a selector (e.g. <c>ControlStyle="color: red;"</c>) are applied to the input itself.
-    /// </summary>
+    /// <inheritdoc cref="IFluentControlStyle.ControlStyle" />
     [Parameter]
     public string? ControlStyle { get; set; }
 
@@ -238,12 +226,11 @@ public partial class FluentTextInput : FluentInputImmediateBase<string?>, IFluen
 
                 await JSRuntime.InvokeVoidAsync("Microsoft.FluentUI.Blazor.Components.TextMasked.applyPatternMask", Id, MaskPattern, MaskLazy, placeholder);
             }
-        }
 
-        if (!string.Equals(ControlStyle, _appliedControlStyle, StringComparison.Ordinal))
-        {
-            _appliedControlStyle = ControlStyle;
-            await JSRuntime.InvokeVoidAsync("Microsoft.FluentUI.Blazor.Components.TextInput.applyControlStyle", Element, ControlStyle);
+            if (!string.IsNullOrEmpty(ControlStyle))
+            {
+                await JSRuntime.InvokeVoidAsync("Microsoft.FluentUI.Blazor.Utilities.Attributes.applyShadowStyle", Element, ".control", ControlStyle);
+            }
         }
     }
